[refactor] #11 CQRS 패턴 적용을 위해 서비스 계층을 Command/Query 기준으로 구조 분리
