+++ src/main/java/org/umc/travlocksserver/domain/auth/service/command/AuthService.java
@@ -1,4 +1,4 @@
-package org.umc.travlocksserver.domain.auth.service;
+package org.umc.travlocksserver.domain.auth.service.command;
 
 import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
